refactor: better file management
