+++ frappe/public/js/frappe/views/desk/widgets/base_widget.js
