+++ frappe/public/js/frappe/views/desk/widgets/index.js
