+++ frappe/public/js/frappe/views/desk/widgets/module_widget.js
